+++ src/features/attend/AttendDateList.tsx
@@ -6,6 +6,7 @@ import AttendDateListElement from './AttendDateListElement';
 import Grid from '@/components/grid';
 import Text from '@/components/text';
 import { mockAttendanceDate, mockMemberAttendance } from '@/mock/attendance';
+import Spacing from '@/components/spacing';
 
 export default function AttendDateList() {
   const [attendanceDates] = useState<string[]>(mockAttendanceDate);
@@ -13,6 +14,7 @@ export default function AttendDateList() {
   return (
     <DefaultPaddedContainer>
       <AttendDateCreation />
+      <Spacing height={20} />
       <hr css={HorizontalLine} />
       <Grid columns={6} css={{ alignItems: 'center', gridTemplateColumns: '1fr 2fr 2fr 2fr 1fr 1fr', padding: '10px 0px' }}>
         <Text css={commonTitleTextStyles}>시작일</Text>
